chore: use production posthog everywhere
It seemed like a good idea at the time to have two different posthog
environments, one for "production" and one for "dev". As we thought more
about what we're doing, though - we don't actually want that. We want to
treat folks who are developing SI as users who are getting a
particular 'package' of SI - in this case, the developer experience!

So we're going to start pushing all the data, including development
data, into the production instance. We can then sort it out later (since
at least now we will have all the data).

Signed-off-by: Adam Jacob <adam@systeminit.com>