feat(web,lang-js,sdf): Support typescript functions
Initial implementation of typescript functions. This PR is ready to be merged, it's stable and the feature generates value for the users, it is limited though. It will require a story to fix the rough edges after the framework with the basic type checking and typescript compiling gets merged.

Then we can parallelize and give more thought on the details, it's hacky and limited, but expansible. We can improve it as needed. This PR is already too big to solve everything.

Adds typescript autocompletions, type checks and suggestions to the frontend. The usability isn't great, I'm sure there is a ton of CodeMirror customizations we could do to have the usability be better. For example it doesn't suggest interface fields if you don't type the first letter. There is no quick hint showing when hovering things like types, etc. Check the comments in this PR to see more details on this.

There are typescript errors that don't relate to any position, but our design doesn't really support this, it may be an easy fix but I didn't want to fall down this rabbithole. We just tie the error to the entire code and let codemirror handle this.

Dynamically generate types for each function, this is a initial implementation, doesn't traverse the graph to generate complex types, it's either scalar, custom function output types (like code-gen, qual, etc), or any. Btw they are generated in a hacky way, we may want to engineer it better, but since it's one moving part of a complex compilation framework I decided to leave the hack there. Some lang-js sandbox types are missing.

Lang-js does not check types properly, for now it's job is to just compile ts to js and executing without caring much. The editor should inform the user of the problems, but the user can still ignore most of it.

I haven't fixed all type errors, but they mostly are related to implicit any, and I don't think we should fix those by adding explicit anys. I have touched way too many functions to give them another thorough look right now, and I don't think that's a must for the initial release.

We should definitely be more strict with typescript, allow the users to customize their settings, or allow them to ignore the checks and just compile blindly to js (which they roughly can right now). Because I'm sure any config will be limited, but I'm more sure that we won't find a middle ground right now, it needs battle testing.