chore(nix) Set up LD_LIBRARY_PATH in development shell environment
This is awful, but necessary (until we find a better way) to be able to
`cargo run` anything that compiles against openssl. Without this, ld is
unable to find libssl.so.3 and libcrypto.so.3.

If we were packaging this up as a flake, instead of only using nix for
the development environment, we'd be using wrapProgram with something
like `--prefix LD_LIBRARY_PATH : ${lib.makeLibraryPath [ openssl ]}`
to make sure the things we're compiling are always using the version of
openssl they were compiled against. Unless we find an even less hacky
way of ensuring the compiled binary is able to find the openssl library
it was compiled against in a nix environment.