fix(dal,sdf): populate child proxies for qualifications
When a qualification is executed at the schema variant context, the
values it creates have to be propagated into the component context for
every component that is an "instance" of that schema variant. We can do
this by creating proxy values at the component context that point back
to the schema variant level values, and executing the functions again to
update the value in the component context. We're not doing this anywhere
automatically for values created by functions run by
`update_from_prototype_function` right now, it's only handled via the
update/insert_for_context PL/pgSQL functions. I've rigged this up here
to do so when a function is executed on the function details panel, but
I believe we should rework this so that `update_from_prototype_function`
handles the population of child proxy values whenever it's called for
automatically. I'm not digging into that yet, since I think it's best
done with a new PL/pgSQL function that will require a little more care
and testing before we're confident it handles every kind of value.