feat(council,pinga) Track failed nodes in DependentValuesUpdate & propagate sub-graph removal through council
If an individual `AttributeValue` fails to update for some reason, we
don't want the entire `DependentValuesUpdate` job to bomb out. Instead,
we want to notify council that we were unable to update that node, so
council can remove the sub-graph with the failed node as the root, and
notify all jobs that are interested in any of the nodes in the sub-graph
that they should no longer be waiting to be told to either start
processing any of those nodes, or to be told that another job has
handled processing one of the nodes.

By not immediately exiting the `DependentValuesUpdate` job, we are also
able to send status updates through to the front end, to let it know
that it should no longer consider any of the failed/removed
`AttributeValue`s as currently updating.