chore(web): fix reactivity issues in code editor
Avoid changing prop array pointer when code is edited. This fixes the
constant reload of the code editor.

Also hotfixes handling of types changing, by reloading the screen. But it
doesn't break usability that much. It should be fixed properly by
reloading ts lang server and retriggering autocomplete/linting, but it's
too much work to unblock the tutorial without much benefit.

It should be done in a subsequent PR

Co-Authored-By: Zack Hamm <zack@systeminit.com>
Co-Authored-By: Paul Stack <zack@systemint.com>
Co-Authored-By: Theo Ephraim <zack@systemint.com>