fix(council,pinga) Reduce race conditions around global graph updates and in-fllight node processing
Given two jobs that are interested in partially overlapping graphs:

```
Job A: AV 1 -> AV 2
Job B: AV 3 -> AV 2
```

There is a race condition that will cause AV 2 to have stale
information, if Job A submits its graph, and is able to start processing
AV 2 before Job B submits its graph, and Job B is able to start and
finish processing AV 3 before Job A has finished processing AV 2.

When council checks to see whether all of the things that AV 2 depends
on have been processed, there is no record that AV 3 was added (and
removed) from the list between when Job A started working on AV 2, and
when it finished, which means that it doesn't know that AV 2 should be
re-evaluated.

To avoid this, we now keep track of both when we determined that Job A
should be the one to process AV 2, and when we modify any of the
informataion about AV 2. If we have modified AV 2's information more
recently than when we determined that Job A should process it, then we
don't consider Job A to have processed the most recent version of AV 2's
inputs. We put Job A's reply channel back at the end of the list of
reply channels that are interested in AV 2, and mark AV 2 as not
currently being processed so it can be picked up as normal when checking
for nodes that are ready to be processed. Job A's reply channel is added
to the back of the list of reply channels interested in AV 2 to try to
more fairly distribute the work of processing nodes that have to be
re-processed like this.

To be able to support re-processing nodes like this, we also make it so
that a job will not remove a node from its list of nodes it is expecting
to be processed until it sees a `BeenProcessed` message from council for
that node, even if it is the one that was doing the processing.