chore: extract common crate deps to use workspace dependencies
This change refactors some of the Rust crate dependency management to
use a Cargo feature called [workspace dependencies]. It may make it
easier to maintain the versions of critical and common crate
dependencies in a more centralized location. While not every crate
should be maintained in this way, the common crates are often rather
obvious, such as tokio, serde, etc.

Additionally, this change also uses a similar features called
[workspace packages]. This should make maintaining certain package level
properties such as `rust-version`, `edition` and `published` in a
more consistent way.

[workspace dependencies]: https://doc.rust-lang.org/cargo/reference/workspaces.html#the-dependencies-table
[workspace packages]: https://doc.rust-lang.org/cargo/reference/workspaces.html#the-package-table

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>