chore(dev_env) Use direnv to load nix devShell out of the box
This .envrc will load any environment variables that are defined in
`.env` (which is ignored by git), and can be used to set up any
developer-specific API keys, or other needed environment variables. It
will also automatically load the default `devShell` defined in the
`flake.nix` to pull in all the dependencies that the repo requires.