feat: lang-js buck2 build
This adds an example of building lang-js with bazel, using custom rules
driving pnpm.

There's a lot happening here that is a little tricky to figure out. It's
compounded by some apparent bugs in the buck2 file detection algorithm
relating to symlinks, which are exacerbated by pnpm using them
extensively.

Signed-off-by: Adam Jacob <adam@systeminit.com>