chore(sdf,si-posthog-rs): remove dependency on OpenSSL
Prior to this change, our current software stack was reliant on the
OpenSSL for via the default feature set of the `reqwest` crate. This
change updates the features for `reqwest` to include the `rustls-tls`
feature which uses the native `rustls` crate instead.

Additionally, this change centralizes the version and configuration of
`reqwest` into the workspace's packages section and finally adds the
`rewest` dependency for building the project with Buck2 (this will be
used in future work).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>