build: build bin/sdf with buck2
This change add Buck2 build and run support to `bin/sdf` so that a
developer can run the following:

```sh
buck2 run //bin/sdf
```

Some refactoring of sdf-server's `config.rs` was needed to determine the
paths to certain keys and files when run under development but these
changes were very similar to the changes needed in veritech-server.

Note that there is still support for running via Cargo (as before).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>