chore(everything) Ensure all enum variants are (and remain) sorted
It's been pretty common to start writing a new enum variant for one of
our (admittedly large) error enums, only to find that the variant you
want already exists with a slightly different name, but you didn't see
it, because the variants are somewhat randomly sorted in many cases.

By adding the `remain` crate, we can annotate our enums such that adding
a new variant that isn't added in sorted order is a compilation error.