feat(*): remove workflows, use actions directly
This PR completely removes workflows and workflow runners.
Recommendations now return an action kind, and that kind is used to find
the action prototype that runs the action recommended, which are run by
the fixes directly. The output from that action is stored on the fix
(and also propagated to the component resource).

This is a breaking change and will require a destructive restart of SI
to run the new migrations.