chore: remove deprecated JWT db code and configuration
This change removes the code and configuration which formerly generated
a JWT signing key that was encrypted in the database. Current behavior
is to let the auth services sign the JWT, leaving the SDF service to
only need to verify a token, meaning that only a known public key is
required.

This refactoring simplifies the configuration required for `//bin/sdf`
at runtime (in development and production), and for testing `//lib/dal`
and `//lib/sdf-server`.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>