build: remove pnpm/Rust crate dependency tracking
This change drops most Rust crates from the pnpm workspace in an effort
to help make Buck2 builds more consistent and reliable.

Prior to this change, calling a `pnpm install` on the workspace created
a `./node_modules` directory for every pnpm workspace member which
included every Rust crate (a `package.json` was generated for each Rust
crate and described the local crate dependencies). Due to the symlinked
nature of pnpm's local `node_modules/` directories and combined with
Buck2's current file-watching symlink-aware de-duplication, many Buck
packages would not be rebuilt when necessary as often a symlink to the
original source was found and used by Buck2 to compute rebuilds.

The change here attempts to--for the moment--preserve the `pnpm
dev:backend`/`pnpm dev:frontend` commands but future work will remove
the entirely as we migrate more fully over to Buck2-driven workflows.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>