chore(pinga) Drastically reduce the default concurrency limit
The default concurrency limit of 50 was apparently dangerously close to
running into exhausting the number of connections in the `PgPool` on
macOS, where the default pool size (based on detected CPU core count) is
significantly lower than the default pool size we've been seeing on our
Linux-based development machines.

This lowers the default concurrency down to something that should be
"safe" on the most common development environments we expect to see.

On systems where the `PgPool` size is significantly higher, and where
system resources generally allow for it, a larger concurrency limit can
be opted in to by setting the environment variable
`SI_PINGA__CONCURRENCY_LIMIT` to something higher. (This can be done
automically by adding it to the `.env`, and `direnv` will handle
loading/updating the env var.)