fix(dal-test): add buck2/cargo awareness when setting up `pkgs_path`
This change slightly alters the configuration setup in `//lib/dal-test`
which is used in `//lib/dal:test-integration` and
`//lib/sdf-server:test-integration`. Prior to this change, when
initially setting up the integration testing we had already declared a
resource dependency on `//pkgs:pkgs` (which is a symlinked folder back
to the source tree's `pkgs/` folder), so the fix was to use this path
value when the test suites run under Buck2 and fall back to using the
`CARGO_MANIFEST_DIR` trick when running under Cargo.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>