feat(sdf) Allow automatic creation & upload of "Workspace Backup" module on change set apply
Whenever a change set is applied, we want to create a module that
contains all of the schemas/schema variants/functions/anything else a
module can contain, and automatically upload it to the module index.
This makes it a bit easier for people to work with things while the
the system is in a high state of flux, and generally makes it easier to
have a backup of any customizations that are made.