feat(web,sdf): create change-set if on head in most actions
Converted most update backend routes to create a change-set if on head.
This uncovered a major problem with this approach though. Which is a
race-condition. If you take two actions fast while on head each action
will happen in a different change-set. The only way to avoid this is
either "freezing" the screen if some update action is happening while on
head, which will unfreeze when the new change set pk is returned, but this
may turn out really awkward for the user as making the unresponsiveness /
breakage of flow clear and ergonomic is hard.

Since this does not affect the regular screen I'm ok with merging and
then we can translate the routes to the approach that is decided.
Reverting this PR is simple enough.

The other option is moving the control of the change set pk back to the front-end.
Which has problems with how we detect change-sets as it requires a route
update, and it may be a dead-end but it would give us total control over those races.
Since JS single-threaded concurrency allows us to intercept all actions.

Restoration actions technically should not be triggearable on head, ignoring
them in this refactor felt awkward as it could show incorrect data on the
screen, so I decided to support them out of the box. They should not
appear, and if one day we decide to show them it will work out of the
box, if not it won't change a thing.