build(buck2): add `check-lint` & `check-type` to ts components
This change adds `check-lint` targets on all TypeScript components in
our system as well as `check-type` targets to verify the TypeScript
types are correct.

The current implementation for the `eslint` rule (which runs our lints)
and the `typescript_check` rule (which runs variants of `tsc --noEmit`)
use the `package_node_modules` pattern we have in `typescript_dist` and
other Node-related rules. The impact here is that each Node sub-package
gets its own isolated `node_modules` and the linting/checking is
happening in a copied directory under `buck-out/`. This is not the most
rapid feedback for developers in their consoles, but likely their code
editors are running plugins to verify eslint and TypeScript warnings as
they go. The current setup should be very remote build friendly, meaning
that remote execution (RE) is less of a lift in future iterations of our
build system.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>