build(buck2): reimpl `check-format` for Rust components
This change updates the `:check-format` targets for Rust components by
using a new `rustfmt_check` rule. This rule does not use a Cargo
subcommand but rather calls `rustfmt` directly meaning that it can check
Rust source formatting for all library and binary crate sources
including examples. It uses a similar technique to the pnpm-related
rules by creating a `__src` output with all Crate source files and
performs a `rustfmt --check` in the output directory. While slightly
slower as a strategy, it should be remote execution-ready, leading to
better concurrency in the near future.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>