build(ci-base): revert Dockerfile to be built on Debian
There is a build correctness issue with `bin/lang-js` which results in a
binary using a standard Linux distro's ELF interpreter and system
libraries. Swapping this image to be based on Alpine Linux (with its
musl libc impl) exposed this correctness issue but before we can land a
fix, reverting the Dockerfile impl seems like a better approach for now.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>