fix(dal): remove list confirmations component view assertion
Remove a ComponentView assertions in the "list_confirmations"
integration test because it is causing intermittent failures. The
intermittent failures are test-specific and removing the assertion does
not impact the goal of the test: to protect against regressions with
listing confirmations.

I'd normally be against removing an assertion in this manner, but given
its low impact, irrelevance to the rest of the test, and likeliness that
it is a bug only experienced within tests (i.e. not in the product
itself), I believe it is not worth keeping in the test.

Signed-off-by: Nick Gerace <nick@systeminit.com>