build(buck2,council): build service Docker images via Buck2/Nix
This change adds Buck2 prelude functionality which enables a developer
to describe a `docker_image` rule in a `BUCK` file that can build a set
of one or more Buck2 targets within a suitable Docker image base. In our
case, we are using our Nix package descriptions for toolchain/build
software and as the initiator of the `buck2 build` process.

The primary work in this change is to support the new `build_dep`
attribute on `docker_image` that takes zero-to-many full qualified Buck2
targets (i.e. `//bin/council:council`, `//app/web:web`, etc.) and can
determine which files and directories need to comprise the resulting
Docker context (i.e. the files and directories uploaded to the Docker
engine for use with the Dockerfile). This is made possible primarily via
a new BXL script whose job it is to return a list of file and directory
paths back to the already-existing `docker_build_context.py` Python
script.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>