fix(dal): commit before applying change-set so jobs finish in correct visibility
Fixes this: https://buildkite.com/system-initiative/si-merge-queue/builds/126#018992fc-533c-4ca1-b22d-37b4be28b237

As the jobs will execute in the change-set's visibility, but they only block after the change-set is merged, their result may endup in a dead change-set and not on main.

Now we block before merging the change-set so we ensure all data from it has been propagated. This also showed a different race on /root/si/name -> /root/domain/name that we hadn't seen any failure for for some reason.

I'm also blocking all commits in the si-macro-test to ensure nothing is ever leaked if changed, it should be a noop for now and a life saver when not a noop anymore.