build(buck2): create a `pnpm_workspace` rule & modify `pnpm_lock`
This change adds a `pnpm_workspace` Buck2 rule which describes the
relationship to a workspace `package.json` and member `package.json`
files. The `pnpm_lock` rule is also updated to be dependent on
`pnpm_workspace` as a pre-requisite dependency. Finally, the `pnpm_lock`
rule no longer attempts to build the lock file itself, but rather uses a
source file in the source tree and relies on external tooling to
generate this file, keep it up-to-date, etc.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>