build(veritech): create Docker image via Buck2/Nix
This turned out to be a **lot** of work. The node `pkg` package would
not produce a working `lang-js` binary when built via the `nix build`
command in the semi-isolated build sandbox for reasons I could not
ultimately fathom.

Instead, the `lang-js` code is essentially run as a `node
path/to/dist/index.js` with carefully installed node modules and a
wrapper script as the "binary" entrypoint. There are several new Buck2
rule types, namely `typescript_runnable_dist` and
`typescript_runnable_dist_bin` where `typescript_runnable_dist` makes
this directory tree of a transpiled Node program with its node_modules
dependencies. This is the primary build artifact which is used when
packaging `lang-js` into a Docker image as a Nix package. However, for
the development experience and integration tests to work, we need a
wrapping script as the entrypoint which can satisfy the
"path-to-lang-js-binary-please" requirement as we formerly shipped this
as a standalone binary executable.

Future consideration should be given to migrating the lang-js
implementation away from a simple Node application to another form which
is easier to package and ship (currently I'd advocate us investigating
Deno for this).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>