feat(auth-api): support sdf tokens in auth-api
We need to make requests to the auth-api from the backend as well as the
frontend. Since we can't re-encode the signed JWT, we only have the one
we've acquired from the auth-api server. This change supports using
tokens in the sdf format for requests to the auth api as well as the
default token format.