Change internal LamportClock implementation to be DateTime<Utc> based instead of Ulid based
(Un?)fortunately, the code is fast enough that we _must_ use the same
`Ulid` generator _everywhere_, to be able to guarantee sorting, as
the time component of a `Ulid` is only precise down to the millisecond.

Because we cannot guarantee that the same `Ulid` generator will be used,
even when operating on the same machine, we now use `DateTime<Utc>` as
the `LamportClock` entry, and serialize/deserialize it as nanoseconds
since the epoch.

This does mean that we lose stable sorting if two entries are generated
within the same unit of precision, but having that go from a millisecond
to a nanosecond should buy us a fair amount of breathing room.