chore(buck2): allow Docker promotions to work with arbitrary tags
This change updates the `docker_image_promote.py` script to accept the
raw tag prefixes needed to take 2 multi-arch image tags of something
like:

- `sha-10aa525707118a4183154c8ca0c2d3bb1bd750db-amd64`
- `sha-10aa525707118a4183154c8ca0c2d3bb1bd750db-arm64v8`

and promote to a multi-arch tag of
`sha-10aa525707118a4183154c8ca0c2d3bb1bd750db`.

Similarily, we can take 2 version-style multi-arch tags like:

- `20230810.211625.0-sha.10aa52570-amd64`
- `20230810.211625.0-sha.10aa52570-arm64v8`

and promote to a multi-arch tag of `20230810.211625.0-sha.10aa52570`.

An added flag of `--update-stable-tag` will now optionally update the
`:stable` tag, but only when asked. This allows us to promote the Git
sha-style tag and the version-style tag as well as update the `:stable`
tag with 2 calls to the script via the `docker_image_promote` Buck2
rule.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>