fix(web,sdf): Ensure exec endpoints always have the latest code
Removes the debounce default in the CodeEditor, but keeps it in case we
feel there's a need for it later. Adds a debounce to the asset code save
editor that mimics the one used to debounce func code saves. Also sends
the entire asset def through the exec endpoint to ensure we save the
latest data immediately prior to executing, ensuring we don't execute
out of date code because of the API debounces.

This ensures we always have the latest code in the API requests to
exec_variant_def and save_and_exec_func.