add --keep-images flag to delete subcommand
When iterating on bits which launch containers as part of the launcher,
the stop/delete/start routine to refresh the containers can take a long
time when downloading the images each time. This change adds a flag and
associated environment variable (SI_KEEP_IMAGES_ON_DELETE) to skip image
cleanup when deleting the containers.

n.b. this change does not provide a way to clean up the images
afterwards, except to start them with `si start` and re-run `si delete`,
or cleaning them manually (e.g., with docker image prune)