fix(dal,sdf-server): clone func when cloning variant def
This commit clones the scaffolding func when cloning the schema variant def to
retain 1:1 relationship constraints.

Without this commit, the func would be re-used from the existing variant
def. Accessor queries will fail when that happens, as observed when
executing "exec_variant_def" after the clone. Specifically, this error
is observed when cloning an asset and immediate creating the clone.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Zachary Hamm <zack@systeminit.com>