fix(si): ensure new binary is in same directory as old before rename
This change fixes an issue which will arise when a user's tempdir is on
a different volume mount than the filesystem mount that contains the
current binary. An attempt to use `fs:rename` (i.e. an atomic move) will
span a filesystem boundary and is often not successful.

A different strategy moves/copies the new binary into the same directory
as the current one, applies the same permissions as the current binary,
and then calls `fs:rename` to move the new binary and replace the old
one. As both files are on the same filesystem, this results in the
desired atomic move.

Thankfully, the [self-replace
crate](https://crates.io/crates/self-replace) performs these exact
activities, so we'll use this!

Closes #2651

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>