Add initial contents of "mostly everything is a node or an edge"
This commits moves the engine-layer towards a design where Workspaces
are snapshots of directed, acyclic graphs. In this design, ChangeSets
point to a existing snapshot rather than overlaying rows on top of each
other in the database. In addition, all graphs are immutable.

This commit also adds scaffolding for the Gobbler, which will handle
processing and performing updates for the new design.

An initial RabbitMQ implementation and library has been added because of
its durable properties, which NATS does not provide.

For data access, this design uses a pull-through cache and a content
store in the database.

Signed-off-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>