Ensure vector clocks are updated when updating node content
Previously, we were relying on `copy_node_index` to also increment the
vector clocks, but we updated `copy_node_index` to be more of a "pure"
copy, which means that it was no longer incrementing the vector clocks.
In `update_content` we explicitly want the vector clocks updated, since
we are doing a write, so we now do this directly.